update offering controller

+++ src/domain/group/entities/group.entity.ts
@@ -45,7 +45,7 @@ export class Group {
   location: string | null;
 
   @ApiProperty({ description: '🈳 class size' })
-  @Column({ type: 'int', unsigned: true, default: 20 })
+  @Column({ type: 'tinyint', unsigned: true, default: 20 })
   capacity: number;
 
   @ApiProperty({ description: '🈳 a comma separated string format' })
